fix: util funcs
