+++ frontend/src/common/utils/utils.ts
@@ -1,0 +1,15 @@
+export function componentToHex(c: number) {
+    const hex = c.toString(16);
+    return hex.length == 1 ? "0" + hex : hex;
+}
+export function hexToRgb(hex: string) {
+    const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+    return result ? [parseInt(result[1], 16), parseInt(result[2], 16), parseInt(result[3], 16)] : null;
+}
+export function rgbToHex(r: number, g: number, b: number) {
+    return componentToHex(r) + componentToHex(g) + componentToHex(b);
+}
+
+export function addVector(a: number[], b: number[]) {
+    return a.map((e, i) => e + b[i]);
+}
